refactor: getDefaultGroupName function
