+++ src/common/modules/index.ts
@@ -50,3 +50,19 @@ export function logger() {
         debug: (...err: any) => console.debug('[DEBUG]:', ...err),
     }
 }
+
+export function getDefaultName(prefix: string): string {
+    const date = new Date()
+    const year = date.getFullYear()
+    const month = addZeroWhenNeeded(date.getMonth() + 1)
+    const day = date.getDate()
+    const hour = addZeroWhenNeeded(date.getHours())
+    const minute = addZeroWhenNeeded(date.getMinutes())
+    const second = addZeroWhenNeeded(date.getSeconds())
+
+    return `${prefix} ${day}.${month}.${year} ${hour}:${minute}:${second}`
+}
+
+function addZeroWhenNeeded(value: number): string {
+    return value < 10 ? `0${value}` : `${value}`
+}
